add unit test for account.sql.go

+++ db/sqlc/querier.go
@@ -10,8 +10,8 @@ import (
 
 type Querier interface {
 	CreateAccount(ctx context.Context, arg CreateAccountParams) (Account, error)
-	DeleteAccount(ctx context.Context, accountNumber int32) error
-	GetAccount(ctx context.Context, accountNumber int32) (Account, error)
+	DeleteAccount(ctx context.Context, accountNumber int64) error
+	GetAccount(ctx context.Context, accountNumber int64) (Account, error)
 	ListAccount(ctx context.Context, arg ListAccountParams) ([]Account, error)
 	UpdateAccount(ctx context.Context, arg UpdateAccountParams) (UpdateAccountRow, error)
 }
